refactor: Move Caddy configuration to separate library
- Extract CaddyManager from internal/server/caddy.go to internal/caddy/manager.go
- Create internal/caddy/types.go with consolidated type definitions
- Update internal/server/server.go imports to use new caddy package
- Remove old internal/server/caddy.go file
- Fix WebSocket support by checking both simple and enhanced services
- Maintain backward compatibility with existing service configurations
- Preserve all functionality: enhanced routing, WebSocket passthrough, HTTP/HTTPS separation

Benefits:
- Separation of concerns: Caddy logic isolated from server logic
- Reusable library: caddy package can be used by other components
- Better maintainability: cleaner codebase structure
- Type safety: proper exports and accessible types

This refactoring maintains zero-trust networking principles and transparent proxy operation.
